pyproject.toml
[project]
name = "deep-learning2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "absl-py>=2.4.0",
    "astunparse>=1.6.3",
    "certifi>=2026.1.4",
    "charset-normalizer>=3.4.4",
    "contourpy>=1.3.3",
    "cupy-cuda12x>=13.6.0",
    "cycler>=0.12.1",
    "flatbuffers>=25.12.19",
    "fonttools>=4.61.1",
    "gast>=0.7.0",
    "google-pasta>=0.2.0",
    "grpcio>=1.78.0",
    "h5py>=3.15.1",
    "idna>=3.11",
    "japanize-matplotlib>=1.1.3",
    "joblib>=1.5.3",
    "kagglehub>=1.0.0",
    "keras>=3.13.2",
    "kiwisolver>=1.4.9",
    "libclang>=18.1.1",
    "markdown>=3.10.2",
    "markdown-it-py>=4.0.0",
    "markupsafe>=3.0.3",
    "matplotlib>=3.10.8",
    "mdurl>=0.1.2",
    "ml-dtypes>=0.5.4",
    "mpmath>=1.3.0",
    "namex>=0.1.0",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "opt-einsum>=3.4.0",
    "optree>=0.18.0",
    "packaging>=26.0",
    "pandas>=3.0.0",
    "pillow>=12.1.1",
    "polars>=1.38.1",
    "protobuf>=6.33.5",
    "pygments>=2.19.2",
    "pyparsing>=3.3.2",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "rich>=14.3.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "six>=1.17.0",
    "sympy>=1.14.0",
    "termcolor>=3.3.0",
    "threadpoolctl>=3.6.0",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "typing-extensions>=4.15.0",
    "tzdata>=2025.3",
    "urllib3>=2.6.3",
    "werkzeug>=3.1.5",
    "wrapt>=2.1.1",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[dependency-groups]
dev = [
    "ruff>=0.15.2",
]
